data-proivider/requirements.txt
pymongo
Flask
gunicorn
requests
flask-cors
pyjwt
attrs